deplay shell

+++ deploy.sh
@@ -1,6 +1,13 @@
 # 删除所有文件(除了public)
-ls |grep -vE 'public|deploy.sh' |xargs rm -r
+d=`pwd`
+echo $d
 
+if [ $d != "/data/go/canbefree.github.io" ]; then 
+    echo "err: please cd $d first"
+    exit
+fi
+
+ls |grep -vE 'public|deploy.sh' |xargs rm -r
 
 mv public/* .
